// web/src/api/models/TypesNftValidateResult.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TypesNftValidateError } from './TypesNftValidateError';
import {
    TypesNftValidateErrorFromJSON,
    TypesNftValidateErrorFromJSONTyped,
    TypesNftValidateErrorToJSON,
    TypesNftValidateErrorToJSONTyped,
} from './TypesNftValidateError';

/**
 * 
 * @export
 * @interface TypesNftValidateResult
 */
export interface TypesNftValidateResult {
    /**
     * 
     * @type {Array<TypesNftValidateError>}
     * @memberof TypesNftValidateResult
     */
    errors?: Array<TypesNftValidateError>;
    /**
     * 
     * @type {boolean}
     * @memberof TypesNftValidateResult
     */
    ok: boolean;
}

/**
 * Check if a given object implements the TypesNftValidateResult interface.
 */
export function instanceOfTypesNftValidateResult(value: object): value is TypesNftValidateResult {
    if (!('ok' in value) || value['ok'] === undefined) return false;
    return true;
}

export function TypesNftValidateResultFromJSON(json: any): TypesNftValidateResult {
    return TypesNftValidateResultFromJSONTyped(json, false);
}

export function TypesNftValidateResultFromJSONTyped(json: any, ignoreDiscriminator: boolean): TypesNftValidateResult {
    if (json == null) {
        return json;
    }
    return {
        
        'errors': json['errors'] == null ? undefined : ((json['errors'] as Array<any>).map(TypesNftValidateErrorFromJSON)),
        'ok': json['ok'],
    };
}

export function TypesNftValidateResultToJSON(json: any): TypesNftValidateResult {
    return TypesNftValidateResultToJSONTyped(json, false);
}

export function TypesNftValidateResultToJSONTyped(value?: TypesNftValidateResult | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'errors': value['errors'] == null ? undefined : ((value['errors'] as Array<any>).map(TypesNftValidateErrorToJSON)),
        'ok': value['ok'],
    };
}