// web/src/api/models/ConfigNftablesConfig.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ConfigNftChain } from './ConfigNftChain';
import {
    ConfigNftChainFromJSON,
    ConfigNftChainFromJSONTyped,
    ConfigNftChainToJSON,
    ConfigNftChainToJSONTyped,
} from './ConfigNftChain';

/**
 * 
 * @export
 * @interface ConfigNftablesConfig
 */
export interface ConfigNftablesConfig {
    /**
     * 
     * @type {{ [key: string]: ConfigNftChain; }}
     * @memberof ConfigNftablesConfig
     */
    chains?: { [key: string]: ConfigNftChain; };
    /**
     * 
     * @type {string}
     * @memberof ConfigNftablesConfig
     */
    defines?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof ConfigNftablesConfig
     */
    include?: Array<string>;
}

/**
 * Check if a given object implements the ConfigNftablesConfig interface.
 */
export function instanceOfConfigNftablesConfig(value: object): value is ConfigNftablesConfig {
    return true;
}

export function ConfigNftablesConfigFromJSON(json: any): ConfigNftablesConfig {
    return ConfigNftablesConfigFromJSONTyped(json, false);
}

export function ConfigNftablesConfigFromJSONTyped(json: any, ignoreDiscriminator: boolean): ConfigNftablesConfig {
    if (json == null) {
        return json;
    }
    return {
        
        'chains': json['chains'] == null ? undefined : (mapValues(json['chains'], ConfigNftChainFromJSON)),
        'defines': json['defines'] == null ? undefined : json['defines'],
        'include': json['include'] == null ? undefined : json['include'],
    };
}

export function ConfigNftablesConfigToJSON(json: any): ConfigNftablesConfig {
    return ConfigNftablesConfigToJSONTyped(json, false);
}

export function ConfigNftablesConfigToJSONTyped(value?: ConfigNftablesConfig | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'chains': value['chains'] == null ? undefined : (mapValues(value['chains'], ConfigNftChainToJSON)),
        'defines': value['defines'],
        'include': value['include'],
    };
}